+++ src/django_rich/test.py
@@ -9,6 +9,7 @@ from typing import TextIO
 from typing import Tuple
 from typing import Type
 from typing import Union
+from typing import cast
 from unittest.case import TestCase
 from unittest.result import STDERR_LINE
 from unittest.result import STDOUT_LINE
@@ -22,6 +23,7 @@ from rich.color import Color
 from rich.console import Console
 from rich.rule import Rule
 from rich.style import Style
+from rich.table import Table
 from rich.traceback import Traceback
 
 _SysExcInfoType = Union[
@@ -161,8 +163,36 @@ class RichPDBDebugResult(PDBDebugResult, RichTextTestResult):
     pass
 
 
-class RichTestRunner(DiscoverRunner.test_runner):  # type: ignore [misc,valid-type]
+class RichTestRunner(unittest.TextTestRunner):
     resultclass = RichTextTestResult
+
+    def _printDurations(self, result: RichTextTestResult) -> None:
+        if not result.collectedDurations:
+            return
+        ls = sorted(result.collectedDurations, key=lambda x: x[1], reverse=True)
+        if cast(int, self.durations) > 0:  # typeshed has a bad hint (?!)
+            ls = ls[: cast(int, self.durations)]
+
+        table = Table(title="Slowest test durations", title_style=YELLOW)
+        table.add_column("Duration", justify="right", no_wrap=True)
+        table.add_column("Test")
+
+        hidden = False
+        for test, elapsed in ls:
+            if self.verbosity < 2 and elapsed < 0.001:
+                hidden = True
+                continue
+            table.add_row("%.3fs" % elapsed, test)
+
+        if table.rows:
+            result.console.print(table)
+        else:
+            result.console.print(table.title, style=table.title_style)
+        if hidden:
+            result.console.print(
+                "\n[i]Durations < 0.001s were hidden. Use -v to show these durations.[/i]",
+                highlight=False,
+            )
 
 
 class RichRunner(DiscoverRunner):