+++ CHANGELOG.rst
@@ -7,6 +7,11 @@ Changelog
 
   Thanks to q0w in `PR #78 <https://github.com/adamchainz/django-rich/pull/78>`__.
 
+* Extend test runner to display test durations from |--durations|__ in a table.
+
+  .. |--durations| replace:: ``--durations``
+  __ https://docs.djangoproject.com/en/stable/ref/django-admin/#cmdoption-test-durations
+
 1.9.0 (2024-06-19)
 ------------------